contrib/intel/jenkins: add loops to main that run all summarizer functions
Add options that correctly iterates through all possible logs that the pipeline
creates for creating a correct functional summary.
This includes every MPI, and every build mode, for the 'all' case and the ability
for case specific selection.
Logs that don't match a specific combination will get skipped in assumption that
the pipeline didn't create any log for that combination

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>