contrib/intel/jenkins: Add a call to envoke the summarizer to jenkinsfile
Enable the summary call in the post clause of the Jenkinsfile.
This cannot be in its own stage because if any of the test stages failed,
then the summary stage will not be run. This is because of the way that
the Jenkins declarative pipeline works.
Users who wish to view the summary will have to view the log output from
the post stage's summary.py call instead of just being able to view the log.
This can maybe be changed in the future, or just have user's be directed to
a website that has this data automatically pushed to it in the future.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>